fix: change input type for Celsius temperature conversion to float

+++ Task_6/Task-3_temp-convert.py
@@ -1,3 +1,3 @@
-celcius = int(input("Введите температуру в градусах Цельсия: "))
+celcius = float(input("Введите температуру в градусах Цельсия: "))
 fahrenheit = (celcius * 9/5) + 32
 print(f"{celcius}°C это {fahrenheit}°F")
